requirements.txt
aiohttp
aiosqlite
pyTelegramBotAPI
python-dotenv
